add login command implement.

+++ AquaMan.WebsocketAdapter/Entity/Money.cs
@@ -6,10 +6,10 @@ namespace AquaMan.WebsocketAdapter.Entity
     public class Money
     {
         public Currency Currency { get; }
-        public ulong Amount { get; private set; }
+        public uint Amount { get; private set; }
         public uint Precise { get;  }
 
-        public Money(Currency currency, ulong amount, uint precise)
+        public Money(Currency currency, uint amount, uint precise)
         {
             Currency = currency;
             Amount = amount;
